Adding the base post clients class
